move glue tests
